requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
decouple==0.0.7
Django==6.0
django-filter==25.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
drf-spectacular==0.29.0
flake8==7.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
librt==0.7.5
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
